JobOutput: extract helper funcs into separate file
